v1.1.9
Bugfixes and sharding for auto reconnection
