+++ package.json
@@ -10,12 +10,13 @@
     "@discordjs/opus": "^0.9.0",
     "@discordjs/voice": "^0.11.0",
     "@napi-rs/canvas": "^0.1.30",
-    "discord.js": "^14.6.0",
+    "discord.js": "^14.7.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "ffmpeg-static": "^5.1.0",
     "glob": "^8.0.3",
     "ora": "^6.1.2",
+    "path": "^0.12.7",
     "play-dl": "^1.9.6",
     "prom-client": "^14.1.0",
     "sequelize": "^6.25.7",
@@ -32,8 +33,5 @@
     "eslint-plugin-prettier": "^4.2.1",
     "tslib": "^2.4.1",
     "typescript": "^5.0.0-dev.20221121"
-  },
-  "overrides": {
-    "discord-api-types": "0.37.14"
   }
 }
